grammars: Add regex checks to the compiler
